pyproject.toml
[tool.poetry]
name = "neo4chat"
version = "0.1.0"
description = "Chat backend app using Neo4j, Redis, FastAPI and Websockets"
authors = ["patacoing <romain.lefebvre1104@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic-settings = "^2.5.2"
neo4j = "^5.25.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
